requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.3.0



